requirements.txt
wheel==0.33.6
Django==1.8
FeinCMS==1.16.2
Markdown==3.1
Pillow==5.4.1
django-mptt==0.9.0
django-suit==0.2.26
django-hosts==1.4